+++ src/component/index.ts
@@ -1,3 +1,4 @@
 export * from "./Anchor";
 export * from "./ErrorPage";
 export * from "./Link";
+export * from "./ThemeToggle";