+++ src/index.ts
@@ -1,2 +1,4 @@
 export * from "./component";
 export * from "./util";
+export * from "./provider";
+export * from "./validation";