+++ src/util/index.ts
@@ -1,4 +1,6 @@
 export * from "./api";
 export * from "./auth";
 export * from "./interface";
+export * from "./session";
+export * from "./theme";
 export * from "./leadingSlash";