+++ src/util/interface/api/APIProp.ts
@@ -1,0 +1,11 @@
+import { API } from "../../api";
+
+/**
+ * A generic property for the {@link API} argument.
+ */
+export interface APIProp {
+    /**
+     * {@link API} instance to make requests with.
+     */
+    "api": API,
+}