+++ src/util/interface/user/GenericUser.ts
@@ -1,0 +1,4 @@
+export interface GenericUser {
+    "name": string,
+    "picture": string
+}